refactored `config` into `tc_cli`
