+++ .github/ISSUE_TEMPLATE/guide.md
@@ -8,20 +8,25 @@ labels: "type: docs"
 <!-- Introduction -->
 
 ## Content Creation Requirements
+
 To maintain the quality of our content, please adhere to the following guidelines:
 
 ### 1. Accuracy in Language
+
 - **Grammar and Spelling:** Ensure your content is free from grammatical and spelling errors.
 
 ### 2. Tone and Style
+
 - **Neutral Tone:** Maintain a neutral, non-emotional tone throughout the content.
 - **Engaging Style:** Write in a free-flowing and engaging manner, keeping the reader's interest.
 
 ### 3. Content Integrity
+
 - **Fact-Checking:** Verify all information. If unsure, use Discord to clarify.
 - **Relevance:** Ensure all content is cohesive, to the point, and directly related to the title.
 
 ### 4. Originality
+
 - **Avoid Low-Effort Content:** Content should be original and not solely generated by AI tools like ChatGPT.
 
 **PS: Adherence to these guidelines is crucial. Content not meeting these standards may require revision or may not be accepted.**
