+++ tc_cli/Cargo.toml
@@ -9,6 +9,8 @@ edition = "2021"
 url = { version = "2", features = ["serde"] }
 serde_yaml = "0.9"
 num_cpus = "1.16.0"
+prost-reflect = { version = "0.12.0", features = ["serde"] }
+
 
 async-graphql-value = "6.0.11"
 cache_control = "0.2.0"
