+++ tc_cli/src/blueprint/from_config/definitions.rs
@@ -3,15 +3,20 @@ use std::collections::BTreeSet;
 use std::hash::Hash;
 
 use regex::Regex;
+use tc_core::blueprint::Type::ListType;
+use tc_core::blueprint::{
+  Cache, Definition, EnumTypeDefinition, EnumValueDefinition, FieldDefinition, GraphQLOperationType,
+  InputFieldDefinition, InputObjectTypeDefinition, InterfaceTypeDefinition, ObjectTypeDefinition, ScalarTypeDefinition,
+  Type, UnionTypeDefinition,
+};
+use tc_core::directive::DirectiveCodec;
+use tc_core::lambda::{Expression, Lambda};
+use tc_core::try_fold::TryFold;
+use tc_core::valid::Valid;
 
-use crate::blueprint::Type::ListType;
 use crate::blueprint::*;
 use crate::config;
-use crate::config::{Config, Field, GraphQLOperationType, Union};
-use crate::directive::DirectiveCodec;
-use crate::lambda::{Expression, Lambda};
-use crate::try_fold::TryFold;
-use crate::valid::Valid;
+use crate::config::{Config, Field, Union};
 
 pub fn to_scalar_type_definition(name: &str) -> Valid<Definition, String> {
   Valid::succeed(Definition::ScalarTypeDefinition(ScalarTypeDefinition {
@@ -278,8 +283,8 @@ fn item_is_numberic(list: &[String]) -> bool {
 
 fn update_resolver_from_path(
   context: &ProcessPathContext,
-  base_field: blueprint::FieldDefinition,
-) -> Valid<blueprint::FieldDefinition, String> {
+  base_field: FieldDefinition,
+) -> Valid<FieldDefinition, String> {
   let has_index = item_is_numberic(context.path);
 
   process_path(context.clone()).and_then(|of_type| {
@@ -361,82 +366,81 @@ fn to_fields(object_name: &str, type_of: &config::Type, config: &Config) -> Vali
     },
   );
 
-  let to_added_field =
-    |add_field: &config::AddField, type_of: &config::Type| -> Valid<blueprint::FieldDefinition, String> {
-      let source_field = type_of
-        .fields
-        .iter()
-        .find(|&(field_name, _)| *field_name == add_field.path[0]);
-      match source_field {
-        Some((_, source_field)) => {
-          let new_field = config::Field {
-            type_of: source_field.type_of.clone(),
-            list: source_field.list,
-            required: source_field.required,
-            list_type_required: source_field.list_type_required,
-            args: source_field.args.clone(),
-            doc: None,
-            modify: source_field.modify.clone(),
-            http: source_field.http.clone(),
-            grpc: source_field.grpc.clone(),
-            unsafe_operation: source_field.unsafe_operation.clone(),
-            const_field: source_field.const_field.clone(),
-            graphql: source_field.graphql.clone(),
-            cache: source_field.cache.clone(),
-          };
-          to_field(&add_field.name, &new_field)
-            .and_then(|field_definition| {
-              let added_field_path = match source_field.http {
-                Some(_) => add_field.path[1..].iter().map(|s| s.to_owned()).collect::<Vec<_>>(),
-                None => add_field.path.clone(),
-              };
-              let invalid_path_handler =
-                |field_name: &str, _added_field_path: &[String], original_path: &[String]| -> Valid<Type, String> {
-                  Valid::fail_with(
-                    "Cannot add field".to_string(),
-                    format!("Path [{}] does not exist", original_path.join(", ")),
-                  )
-                  .trace(field_name)
-                };
-              let path_resolver_error_handler = |resolver_name: &str,
-                                                 field_type: &str,
-                                                 field_name: &str,
-                                                 original_path: &[String]|
-               -> Valid<Type, String> {
-                Valid::<Type, String>::fail_with(
+  let to_added_field = |add_field: &config::AddField, type_of: &config::Type| -> Valid<FieldDefinition, String> {
+    let source_field = type_of
+      .fields
+      .iter()
+      .find(|&(field_name, _)| *field_name == add_field.path[0]);
+    match source_field {
+      Some((_, source_field)) => {
+        let new_field = config::Field {
+          type_of: source_field.type_of.clone(),
+          list: source_field.list,
+          required: source_field.required,
+          list_type_required: source_field.list_type_required,
+          args: source_field.args.clone(),
+          doc: None,
+          modify: source_field.modify.clone(),
+          http: source_field.http.clone(),
+          grpc: source_field.grpc.clone(),
+          unsafe_operation: source_field.unsafe_operation.clone(),
+          const_field: source_field.const_field.clone(),
+          graphql: source_field.graphql.clone(),
+          cache: source_field.cache.clone(),
+        };
+        to_field(&add_field.name, &new_field)
+          .and_then(|field_definition| {
+            let added_field_path = match source_field.http {
+              Some(_) => add_field.path[1..].iter().map(|s| s.to_owned()).collect::<Vec<_>>(),
+              None => add_field.path.clone(),
+            };
+            let invalid_path_handler =
+              |field_name: &str, _added_field_path: &[String], original_path: &[String]| -> Valid<Type, String> {
+                Valid::fail_with(
                   "Cannot add field".to_string(),
-                  format!(
-                    "Path: [{}] contains resolver {} at [{}.{}]",
-                    original_path.join(", "),
-                    resolver_name,
-                    field_type,
-                    field_name
-                  ),
+                  format!("Path [{}] does not exist", original_path.join(", ")),
                 )
+                .trace(field_name)
               };
-              update_resolver_from_path(
-                &ProcessPathContext {
-                  path: &added_field_path,
-                  field: source_field,
-                  type_info: type_of,
-                  is_required: false,
-                  config,
-                  invalid_path_handler: &invalid_path_handler,
-                  path_resolver_error_handler: &path_resolver_error_handler,
-                  original_path: &add_field.path,
-                },
-                field_definition,
+            let path_resolver_error_handler = |resolver_name: &str,
+                                               field_type: &str,
+                                               field_name: &str,
+                                               original_path: &[String]|
+             -> Valid<Type, String> {
+              Valid::<Type, String>::fail_with(
+                "Cannot add field".to_string(),
+                format!(
+                  "Path: [{}] contains resolver {} at [{}.{}]",
+                  original_path.join(", "),
+                  resolver_name,
+                  field_type,
+                  field_name
+                ),
               )
-            })
-            .trace(config::AddField::trace_name().as_str())
-        }
-        None => Valid::fail(format!(
-          "Could not find field {} in path {}",
-          add_field.path[0],
-          add_field.path.join(",")
-        )),
+            };
+            update_resolver_from_path(
+              &ProcessPathContext {
+                path: &added_field_path,
+                field: source_field,
+                type_info: type_of,
+                is_required: false,
+                config,
+                invalid_path_handler: &invalid_path_handler,
+                path_resolver_error_handler: &path_resolver_error_handler,
+                original_path: &add_field.path,
+              },
+              field_definition,
+            )
+          })
+          .trace(config::AddField::trace_name().as_str())
       }
-    };
+      None => Valid::fail(format!(
+        "Could not find field {} in path {}",
+        add_field.path[0],
+        add_field.path.join(",")
+      )),
+    }
+  };
 
   let added_fields = Valid::from_iter(type_of.added_fields.iter(), |added_field| {
     to_added_field(added_field, type_of)
