+++ tc_cli/src/blueprint/from_config/operators/graphql.rs
@@ -1,10 +1,11 @@
-use crate::blueprint::FieldDefinition;
-use crate::config::{self, Config, Field, GraphQLOperationType};
-use crate::graphql::RequestTemplate;
-use crate::helpers;
-use crate::lambda::Lambda;
-use crate::try_fold::TryFold;
-use crate::valid::{Valid, ValidationError};
+use tc_core::blueprint::{FieldDefinition, GraphQLOperationType};
+use tc_core::graphql::RequestTemplate;
+use tc_core::helpers;
+use tc_core::lambda::Lambda;
+use tc_core::try_fold::TryFold;
+use tc_core::valid::{Valid, ValidationError};
+
+use crate::config::{self, Config, Field};
 
 pub fn update_graphql<'a>(
   operation_type: &'a GraphQLOperationType,
