+++ tc_cli/src/blueprint/from_config/operators/http.rs
@@ -1,13 +1,16 @@
+use tc_core::blueprint::group_by::GroupBy;
+use tc_core::blueprint::{FieldDefinition, Type};
+use tc_core::endpoint::Endpoint;
+use tc_core::helpers;
+use tc_core::http::{Method, RequestTemplate};
+use tc_core::lambda::{Expression, Lambda, Unsafe};
+use tc_core::try_fold::TryFold;
+use tc_core::valid::{Valid, ValidationError};
+
 use crate::blueprint::from_config::to_type;
 use crate::blueprint::*;
-use crate::config::group_by::GroupBy;
+use crate::config;
 use crate::config::{Config, Field};
-use crate::endpoint::Endpoint;
-use crate::http::{Method, RequestTemplate};
-use crate::lambda::{Expression, Lambda, Unsafe};
-use crate::try_fold::TryFold;
-use crate::valid::{Valid, ValidationError};
-use crate::{config, helpers};
 
 struct MustachePartsValidator<'a> {
   type_of: &'a config::Type,
