+++ tc_cli/src/blueprint/from_config/operators/modify.rs
@@ -1,9 +1,10 @@
-use crate::blueprint::*;
+use tc_core::blueprint::FieldDefinition;
+use tc_core::lambda::Lambda;
+use tc_core::try_fold::TryFold;
+use tc_core::valid::Valid;
+
 use crate::config;
 use crate::config::{Config, Field};
-use crate::lambda::Lambda;
-use crate::try_fold::TryFold;
-use crate::valid::Valid;
 
 pub fn update_modify<'a>() -> TryFold<'a, (&'a Config, &'a Field, &'a config::Type, &'a str), FieldDefinition, String> {
   TryFold::<(&Config, &Field, &config::Type, &'a str), FieldDefinition, String>::new(
