+++ tc_cli/src/blueprint/from_config/operators/unsafe_field.rs
@@ -1,9 +1,10 @@
-use crate::blueprint::*;
+use tc_core::blueprint::FieldDefinition;
+use tc_core::lambda::Lambda;
+use tc_core::try_fold::TryFold;
+use tc_core::valid::Valid;
+
 use crate::config;
 use crate::config::{Config, Field};
-use crate::lambda::Lambda;
-use crate::try_fold::TryFold;
-use crate::valid::Valid;
 
 pub fn update_unsafe<'a>() -> TryFold<'a, (&'a Config, &'a Field, &'a config::Type, &'a str), FieldDefinition, String> {
   TryFold::<(&Config, &Field, &config::Type, &str), FieldDefinition, String>::new(|(_, field, _, _), b_field| {
