+++ tc_cli/src/cli/tc.rs
@@ -7,13 +7,13 @@ use inquire::Confirm;
 use resource::resource_str;
 use stripmargin::StripMargin;
 use tc_core::blueprint::Blueprint;
-use tc_core::config::Config;
 use tc_core::print_schema;
 use tokio::runtime::Builder;
 
 use super::command::{Cli, Command};
 use crate::cli::fmt::Fmt;
 use crate::cli::CLIError;
+use crate::config::Config;
 use crate::http::server::Server;
 
 pub fn run() -> Result<()> {
