+++ tc_cli/src/config/server.rs
@@ -1,9 +1,11 @@
-use std::collections::{BTreeMap, BTreeSet};
+use std::collections::BTreeMap;
+use std::net::{AddrParseError, IpAddr};
 
-use derive_setters::Setters;
+use hyper::header::{HeaderName, HeaderValue};
+use hyper::HeaderMap;
 use serde::{Deserialize, Serialize};
-
-use crate::config::{is_default, KeyValues};
+use tc_core::blueprint::{is_default, Http, HttpVersion, KeyValues};
+use tc_core::valid::{Valid, ValidationError};
 
 #[derive(Serialize, Deserialize, Clone, Debug, Default, PartialEq, Eq)]
 #[serde(rename_all = "camelCase")]
@@ -42,26 +44,6 @@ pub struct Server {
   pub key: Option<String>,
   #[serde(default, skip_serializing_if = "is_default")]
   pub pipeline_flush: Option<bool>,
-}
-
-#[derive(Deserialize, Serialize, Debug, PartialEq, Eq, Clone, Default)]
-pub enum HttpVersion {
-  #[default]
-  HTTP1,
-  HTTP2,
-}
-
-#[derive(Serialize, Deserialize, PartialEq, Eq, Clone, Debug, Setters)]
-#[serde(rename_all = "camelCase", default)]
-pub struct Batch {
-  pub max_size: usize,
-  pub delay: usize,
-  pub headers: BTreeSet<String>,
-}
-impl Default for Batch {
-  fn default() -> Self {
-    Batch { max_size: 100, delay: 1, headers: BTreeSet::new() }
-  }
 }
 
 impl Server {
@@ -144,114 +126,85 @@ impl Server {
   }
 }
 
-#[derive(Serialize, Deserialize, PartialEq, Eq, Clone, Debug)]
-pub struct Proxy {
-  pub url: String,
-}
+impl TryInto<tc_core::blueprint::Server> for Server {
+  type Error = ValidationError<String>;
 
-#[derive(Serialize, Deserialize, PartialEq, Eq, Clone, Debug, Setters, Default)]
-#[serde(rename_all = "camelCase", default)]
-pub struct Upstream {
-  #[serde(default, skip_serializing_if = "is_default")]
-  pub pool_idle_timeout: Option<u64>,
-  #[serde(default, skip_serializing_if = "is_default")]
-  pub pool_max_idle_per_host: Option<usize>,
-  #[serde(default, skip_serializing_if = "is_default")]
-  pub keep_alive_interval: Option<u64>,
-  #[serde(default, skip_serializing_if = "is_default")]
-  pub keep_alive_timeout: Option<u64>,
-  #[serde(default, skip_serializing_if = "is_default")]
-  pub keep_alive_while_idle: Option<bool>,
-  #[serde(default, skip_serializing_if = "is_default")]
-  pub proxy: Option<Proxy>,
-  #[serde(default, skip_serializing_if = "is_default")]
-  pub connect_timeout: Option<u64>,
-  #[serde(default, skip_serializing_if = "is_default")]
-  pub timeout: Option<u64>,
-  #[serde(default, skip_serializing_if = "is_default")]
-  pub tcp_keep_alive: Option<u64>,
-  #[serde(default, skip_serializing_if = "is_default")]
-  pub user_agent: Option<String>,
-  #[serde(default, skip_serializing_if = "is_default")]
-  pub allowed_headers: Option<BTreeSet<String>>,
-  #[serde(rename = "baseURL", default, skip_serializing_if = "is_default")]
-  pub base_url: Option<String>,
-  #[serde(default, skip_serializing_if = "is_default")]
-  pub http_cache: Option<bool>,
-  #[serde(default, skip_serializing_if = "is_default")]
-  pub batch: Option<Batch>,
-}
+  fn try_into(self) -> Result<tc_core::blueprint::Server, Self::Error> {
+    let http_server = match self.clone().get_version() {
+      HttpVersion::HTTP2 => {
+        let cert = Valid::from_option(self.cert.clone(), "Certificate is required for HTTP2".to_string());
+        let key = Valid::from_option(self.key.clone(), "Key is required for HTTP2".to_string());
 
-impl Upstream {
-  pub fn get_pool_idle_timeout(&self) -> u64 {
-    self.pool_idle_timeout.unwrap_or(60)
-  }
-  pub fn get_pool_max_idle_per_host(&self) -> usize {
-    self.pool_max_idle_per_host.unwrap_or(60)
-  }
-  pub fn get_keep_alive_interval(&self) -> u64 {
-    self.keep_alive_interval.unwrap_or(60)
-  }
-  pub fn get_keep_alive_timeout(&self) -> u64 {
-    self.keep_alive_timeout.unwrap_or(60)
-  }
-  pub fn get_keep_alive_while_idle(&self) -> bool {
-    self.keep_alive_while_idle.unwrap_or(false)
-  }
-  pub fn get_connect_timeout(&self) -> u64 {
-    self.connect_timeout.unwrap_or(60)
-  }
-  pub fn get_timeout(&self) -> u64 {
-    self.timeout.unwrap_or(60)
-  }
-  pub fn get_tcp_keep_alive(&self) -> u64 {
-    self.tcp_keep_alive.unwrap_or(5)
-  }
-  pub fn get_user_agent(&self) -> String {
-    self.user_agent.clone().unwrap_or("Tailcall/1.0".to_string())
-  }
-  pub fn get_enable_http_cache(&self) -> bool {
-    self.http_cache.unwrap_or(false)
-  }
-  pub fn get_allowed_headers(&self) -> BTreeSet<String> {
-    self.allowed_headers.clone().unwrap_or_default()
-  }
-  pub fn get_delay(&self) -> usize {
-    self.batch.clone().unwrap_or_default().delay
-  }
-
-  pub fn get_max_size(&self) -> usize {
-    self.batch.clone().unwrap_or_default().max_size
-  }
-
-  pub fn merge_right(mut self, other: Self) -> Self {
-    self.allowed_headers = other.allowed_headers.map(|other| {
-      if let Some(mut self_headers) = self.allowed_headers {
-        self_headers.extend(&mut other.iter().map(|s| s.to_owned()));
-        self_headers
-      } else {
-        other
+        cert.zip(key).map(|(cert, key)| Http::HTTP2 { cert, key })
       }
-    });
-    self.base_url = other.base_url.or(self.base_url);
-    self.connect_timeout = other.connect_timeout.or(self.connect_timeout);
-    self.http_cache = other.http_cache.or(self.http_cache);
-    self.keep_alive_interval = other.keep_alive_interval.or(self.keep_alive_interval);
-    self.keep_alive_timeout = other.keep_alive_timeout.or(self.keep_alive_timeout);
-    self.keep_alive_while_idle = other.keep_alive_while_idle.or(self.keep_alive_while_idle);
-    self.pool_idle_timeout = other.pool_idle_timeout.or(self.pool_idle_timeout);
-    self.pool_max_idle_per_host = other.pool_max_idle_per_host.or(self.pool_max_idle_per_host);
-    self.proxy = other.proxy.or(self.proxy);
-    self.tcp_keep_alive = other.tcp_keep_alive.or(self.tcp_keep_alive);
-    self.timeout = other.timeout.or(self.timeout);
-    self.user_agent = other.user_agent.or(self.user_agent);
-    self.batch = other.batch.map(|other| {
-      let mut batch = self.batch.unwrap_or_default();
-      batch.max_size = other.max_size;
-      batch.delay = other.delay;
-      batch.headers.extend(other.headers);
-      batch
-    });
-    self
+      _ => Valid::succeed(Http::HTTP1),
+    };
+
+    validate_hostname((self).get_hostname().to_lowercase())
+      .zip(http_server)
+      .zip(handle_response_headers((self).get_response_headers().0))
+      .map(|((hostname, http), response_headers)| tc_core::blueprint::Server {
+        enable_apollo_tracing: (self).enable_apollo_tracing(),
+        enable_cache_control_header: (self).enable_cache_control(),
+        enable_graphiql: (self).enable_graphiql(),
+        enable_introspection: (self).enable_introspection(),
+        enable_query_validation: (self).enable_query_validation(),
+        enable_response_validation: (self).enable_http_validation(),
+        enable_batch_requests: (self).enable_batch_requests(),
+        global_response_timeout: (self).get_global_response_timeout(),
+        http,
+        worker: (self).get_workers(),
+        port: (self).get_port(),
+        hostname,
+        vars: (self).get_vars(),
+        pipeline_flush: (self).get_pipeline_flush(),
+        response_headers,
+      })
+      .to_result()
+  }
+}
+
+fn validate_hostname(hostname: String) -> Valid<IpAddr, String> {
+  if hostname == "localhost" {
+    Valid::succeed(IpAddr::from([127, 0, 0, 1]))
+  } else {
+    Valid::from(
+      hostname
+        .parse()
+        .map_err(|e: AddrParseError| ValidationError::new(format!("Parsing failed because of {}", e))),
+    )
+    .trace("hostname")
+    .trace("@server")
+    .trace("schema")
+  }
+}
+
+fn handle_response_headers(resp_headers: BTreeMap<String, String>) -> Valid<HeaderMap, String> {
+  Valid::from_iter(resp_headers.iter(), |(k, v)| {
+    let name = Valid::from(
+      HeaderName::from_bytes(k.as_bytes())
+        .map_err(|e| ValidationError::new(format!("Parsing failed because of {}", e))),
+    );
+    let value = Valid::from(
+      HeaderValue::from_str(v.as_str()).map_err(|e| ValidationError::new(format!("Parsing failed because of {}", e))),
+    );
+    name.zip(value)
+  })
+  .map(|headers| headers.into_iter().collect::<HeaderMap>())
+  .trace("responseHeaders")
+  .trace("@server")
+  .trace("schema")
+}
+
+#[cfg(test)]
+mod tests {
+  use tc_core::valid::ValidationError;
+
+  use crate::config;
+
+  #[test]
+  fn test_try_from_default() {
+    let actual: Result<tc_core::blueprint::Server, ValidationError<String>> = config::Server::default().try_into();
+    assert!(actual.is_ok())
   }
 }
