+++ tc_cli/src/http/server.rs
@@ -2,13 +2,13 @@ use std::sync::Arc;
 
 use anyhow::Result;
 use tc_core::blueprint::{Blueprint, Http};
-use tc_core::config::Config;
 use tokio::sync::oneshot::{self};
 
 use super::http_1::start_http_1;
 use super::http_2::start_http_2;
 use super::server_config::ServerConfig;
 use crate::cli::CLIError;
+use crate::config::Config;
 
 pub struct Server {
   config: Config,
