+++ tc_cli/src/lib.rs
@@ -1,4 +1,6 @@
-pub mod cli;
-pub mod http;
-pub mod config;
+#![allow(clippy::module_inception)]
+
 pub mod blueprint;
+pub mod cli;
+pub mod config;
+pub mod http;
