+++ tc_core/src/blueprint/blueprint.rs
@@ -8,9 +8,8 @@ use async_graphql::*;
 use derive_setters::Setters;
 use serde_json::Value;
 
-use super::GlobalTimeout;
-use crate::blueprint::from_config::Server;
-use crate::config::Upstream;
+use super::{GlobalTimeout, Upstream};
+use crate::blueprint::Server;
 use crate::lambda::{Expression, Lambda};
 
 /// Blueprint is an intermediary representation that allows us to generate graphQL APIs.
