+++ tc_core/src/blueprint/mod.rs
@@ -1,11 +1,14 @@
 mod blueprint;
-mod compress;
+pub mod compress;
 mod const_utils;
+pub mod for_config;
 mod into_schema;
 mod timeout;
-mod server;
-
 pub use blueprint::*;
 pub use const_utils::*;
-pub use server::*;
+pub use for_config::*;
 pub use timeout::GlobalTimeout;
+
+pub fn is_default<T: Default + Eq>(val: &T) -> bool {
+  *val == T::default()
+}
