+++ tc_core/src/grpc/data_loader.rs
@@ -11,8 +11,8 @@ use async_graphql_value::ConstValue;
 use super::data_loader_request::DataLoaderRequest;
 use super::protobuf::ProtobufOperation;
 use super::request::execute_grpc_request;
-use crate::config::group_by::GroupBy;
-use crate::config::Batch;
+use crate::blueprint::group_by::GroupBy;
+use crate::blueprint::Batch;
 use crate::data_loader::{DataLoader, Loader};
 use crate::grpc::request::create_grpc_request;
 use crate::http::{HttpClient, Response};
