+++ tests/graphql_spec.rs
@@ -13,8 +13,8 @@ use pretty_assertions::{assert_eq, assert_ne};
 use regex::Regex;
 use serde::{Deserialize, Serialize};
 use serde_json::Value;
+use tc_cli::config::Config;
 use tc_core::blueprint::Blueprint;
-use tc_core::config::Config;
 use tc_core::directive::DirectiveCodec;
 use tc_core::http::server_context::ServerContext;
 use tc_core::http::RequestContext;
