+++ tests/http_spec.rs
@@ -16,9 +16,9 @@ use pretty_assertions::assert_eq;
 use reqwest::header::{HeaderName, HeaderValue};
 use serde::{Deserialize, Serialize};
 use serde_json::Value;
+use tc_cli::config::{Config, Source};
 use tc_core::async_graphql_hyper::{GraphQLBatchRequest, GraphQLRequest};
 use tc_core::blueprint::Blueprint;
-use tc_core::config::{Config, Source};
 use tc_core::http::server_context::ServerContext;
 use tc_core::http::{handle_request, HttpClient, Method, Response};
 use url::Url;
