+++ tests/server_spec.rs
@@ -1,7 +1,7 @@
 use reqwest::Client;
 use serde_json::json;
+use tc_cli::config::Config;
 use tc_cli::http::server::Server;
-use tc_core::config::Config;
 
 async fn test_server(configs: &[&str], url: &str) {
   let config = Config::read_from_files(configs.iter()).await.unwrap();
